Support deriving for data family instances
`deriveLift` and `makeLift` can now derive for data family instances if you pass them the name of a data family instance's constructor. Note that this feature only works with GHC 7.4 and above, since that is when Template Haskell was given the ability to reify data family Info.

This fixes #17 once and for all.