Remove Lift (), Ratio a, Identity a and Const a
This is a fix of a mess I started. The idea is to separate
the three packages responsibilities:
- `th-lift` provides `TemplateHaskell` based deriving code for `Lift`
  instances (and defines `Lift Name`)
- `th-orphans` provides rest of `Lift` instances for
  `Language.Haskell.TH` namespace
- `th-lift-instances` provides instances for small set of core packages,
  and acts as a compat module (e.g. provides `Lift ()`).

Note: `th-lift-instances` uses `TemplateHaskellQuotes` only nowadays.
(Which s sometimes handy, e.g. when splicing TemplateHaskell is broken...)
That's why I removed `th-lift` dependency, and didn't notice the
overlapping orphan instances problem in the first place.