Allow building with template-haskell-2.17.* (#45)
To minimize the amount of CPP that I had to use, I copied over
`unsafeSpliceCoerce` from `th-compat` and put it in a new
`Language.Haskell.TH.Lift.Internal` module. Since
`unsafeSpliceCoerce` appears in generated code, I decided to export
`Language.Haskell.TH.Lift.Internal`, but with an explicit caveat that
it is not a part of the public API.

While I was in town, I also took the opportunity to fix #43.